requirements.txt
openai
flask
gunicorn
flask-cors
python-dotenv
chromadb>=0.4.24
sentence-transformers
tiktoken
rapidfuzz
